requirements.txt
python-dotenv>=1.0.0
mysql-connector-python>=2.0.0
psycopg2-binary>=2.0.0
pytest>=6.0.0
